Restructuring package to follow ros naming conventions
