+++ strands_sweep_registration/CMakeLists.txt
@@ -16,13 +16,15 @@ find_package(catkin REQUIRED COMPONENTS
 rosbuild_prepare_qt4(QtCore QtXml)
 
 catkin_package(
-   INCLUDE_DIRS src src/Camera src/Camera/PixelFunction src/Frame src/Sweep
+   INCLUDE_DIRS include
    CATKIN_DEPENDS metaroom_xml_parser image_geometry cpp_common roscpp rosconsole tf_conversions
    DEPENDS Ceres
    LIBRARIES PixelFunction_strands_room Camera_strands_room  Frame_strands_room Sweep_strands_room util_strands_room pair3DError_strands_room ProblemFrameConnection_strands_room RobotContainer_strands_room
 )
 
-include_directories(${catkin_INCLUDE_DIRS})
+include_directories(${catkin_INCLUDE_DIRS}
+                    include
+                    )
 
 # PCL
 find_package(PCL 1.7 REQUIRED)
@@ -45,39 +47,31 @@ link_directories(/opt/ros/indigo/lib)
 FIND_PACKAGE(Ceres REQUIRED)
 INCLUDE_DIRECTORIES(${CERES_INCLUDE_DIRS})
 
-include_directories(src)
-include_directories(src/Camera)
-include_directories(src/Camera/PixelFunction)
-include_directories(src/Frame)
-include_directories(src/Match)
-include_directories(src/PoseTransform)
-include_directories(src/Sweep)
-
 
 
 #PixelFunction is a per pixel library for updating the pixelvalues individually. Will learn per pixel biases and fix those
-add_library(PixelFunction_strands_room src/Camera/PixelFunction/PixelFunction.cpp)
+add_library(PixelFunction_strands_room include/strands_sweep_registration/PixelFunction.h src/PixelFunction.cpp)
 
 #Keeps camera parameters etc
-add_library(Camera_strands_room src/Camera/Camera.cpp)
+add_library(Camera_strands_room include/strands_sweep_registration/Camera.h src/Camera.cpp)
 target_link_libraries(Camera_strands_room PixelFunction_strands_room ${QT_QTMAIN_LIBRARY} ${QT_LIBRARIES} ${catkin_LIBRARIES} ${PCL_LIBRARIES} ${CERES_LIBRARIES})
 
-add_library(Frame_strands_room src/Frame/Frame.cpp)
+add_library(Frame_strands_room include/strands_sweep_registration/Frame.h src/Frame.cpp)
 target_link_libraries(Frame_strands_room Camera_strands_room  ${QT_QTMAIN_LIBRARY} ${QT_LIBRARIES} ${catkin_LIBRARIES} ${PCL_LIBRARIES} ${CERES_LIBRARIES})
 
-add_library(Sweep_strands_room src/Sweep/Sweep.cpp)
+add_library(Sweep_strands_room include/strands_sweep_registration/Sweep.h src/Sweep.cpp)
 target_link_libraries(Sweep_strands_room Frame_strands_room ${QT_QTMAIN_LIBRARY} ${QT_LIBRARIES} ${catkin_LIBRARIES} ${PCL_LIBRARIES} ${CERES_LIBRARIES})
 
-add_library(util_strands_room src/util.cpp)
+add_library(util_strands_room include/strands_sweep_registration/util.h src/util.cpp)
 target_link_libraries(util_strands_room ${QT_QTMAIN_LIBRARY} ${QT_LIBRARIES} ${catkin_LIBRARIES} ${PCL_LIBRARIES} ${CERES_LIBRARIES})
 
-add_library(pair3DError_strands_room src/pair3DError.cpp)
+add_library(pair3DError_strands_room include/strands_sweep_registration/pair3DError.h src/pair3DError.cpp)
 target_link_libraries(pair3DError_strands_room util_strands_room ${QT_QTMAIN_LIBRARY} ${QT_LIBRARIES} ${catkin_LIBRARIES} ${PCL_LIBRARIES} ${CERES_LIBRARIES})
 
-add_library(ProblemFrameConnection_strands_room src/ProblemFrameConnection.cpp)
+add_library(ProblemFrameConnection_strands_room include/strands_sweep_registration/ProblemFrameConnection.h src/ProblemFrameConnection.cpp)
 target_link_libraries(ProblemFrameConnection_strands_room pair3DError_strands_room Frame_strands_room ${QT_QTMAIN_LIBRARY} ${QT_LIBRARIES} ${catkin_LIBRARIES} ${PCL_LIBRARIES} ${CERES_LIBRARIES})
 
-add_library(RobotContainer_strands_room src/RobotContainer.cpp)
+add_library(RobotContainer_strands_room include/strands_sweep_registration/RobotContainer.h src/RobotContainer.cpp)
 target_link_libraries(RobotContainer_strands_room ProblemFrameConnection_strands_room pair3DError_strands_room util_strands_room Sweep_strands_room ${QT_QTMAIN_LIBRARY} ${QT_LIBRARIES} ${PCL_LIBRARIES} ${CERES_LIBRARIES})
 
 ############################# INSTALL TARGETS
@@ -88,13 +82,7 @@ install(TARGETS PixelFunction_strands_room Camera_strands_room  Frame_strands_ro
   RUNTIME DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
 )
 
-install(DIRECTORY include/semantic_map/
+install(DIRECTORY include/strands_sweep_registration/
   DESTINATION ${CATKIN_PACKAGE_INCLUDE_DESTINATION}
 )
-
-
-install(DIRECTORY launch/
-  DESTINATION ${CATKIN_PACKAGE_SHARE_DESTINATION}/launch
-)
-
