+++ strands_sweep_registration/include/strands_sweep_registration/Camera.h
@@ -5,7 +5,7 @@
 #include <string>
 #include <stdio.h>
 
-#include "PixelFunction/PixelFunction.h"
+#include "PixelFunction.h"
 
 class Camera {
 	public:
