+++ strands_sweep_registration/src/RobotContainer.cpp
@@ -1,4 +1,4 @@
-#include "RobotContainer.h"
+#include "strands_sweep_registration/RobotContainer.h"
 
 //#include "simple_xml_parser.h"
 //#include "simple_summary_parser.h"
@@ -6,7 +6,7 @@
 #include <tf_conversions/tf_eigen.h>
 #include <load_utilities.h>
 
-#include "pair3DError.h"
+#include "strands_sweep_registration/pair3DError.h"
 
 typedef pcl::PointXYZRGB PointType;
 typedef typename SimpleSummaryParser::EntityStruct Entities;
